server/requirements.txt
fastapi==0.136.0
httpx==0.28.1
langchain_classic==1.0.4
langchain_core==1.3.0
langchain_google_genai==4.2.2
numpy==2.2.6
pgvector==0.4.2
pydantic==2.13.3
jwt==1.4.0
python-dotenv==1.2.2
bcrypt
Requests==2.33.1
SQLAlchemy==2.0.49
uvicorn
python-multipart
fastapi[standard]                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               
psycopg2-binary
tzdata
httpx[http2]